math: Fix UB in test-canonical-ldbl-96.c
UBSAN: Undefined behaviour in ../sysdeps/ieee754/ldbl-96/test-canonical-ldbl-96.c:97:7 left shift of 1 by 31 cannot be represented in type 'int'

+++ sysdeps/ieee754/ldbl-96/test-canonical-ldbl-96.c
@@ -95,7 +95,8 @@ do_test (void)
     {
       long double ld;
       SET_LDOUBLE_WORDS (ld, tests[i].exponent | (tests[i].sign << 15),
-			 (tests[i].mantissa >> 32) | (tests[i].high << 31),
+			 (tests[i].mantissa >> 32)
+			 | ((uint32_t)tests[i].high << 31),
 			 tests[i].mantissa & 0xffffffffULL);
       bool canonical = iscanonical (ld);
       if (canonical == tests[i].canonical)
